General refactoring of code
