+++ src/test/java/com/adobe/campaign/tests/logparser/core/LogDataTest.java
@@ -9,7 +9,7 @@
  *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.adobe.campaign.tests.logparser;
+package com.adobe.campaign.tests.logparser.core;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
@@ -18,6 +18,7 @@ import static org.testng.Assert.assertThrows;
 import java.io.File;
 import java.util.*;
 
+import com.adobe.campaign.tests.logparser.core.*;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.filefilter.TrueFileFilter;
 import org.apache.commons.io.filefilter.WildcardFileFilter;
@@ -25,10 +26,6 @@ import org.hamcrest.Matchers;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
-import com.adobe.campaign.tests.logparser.LogData;
-import com.adobe.campaign.tests.logparser.GenericEntry;
-import com.adobe.campaign.tests.logparser.ParseDefinition;
-import com.adobe.campaign.tests.logparser.ParseDefinitionEntry;
 import com.adobe.campaign.tests.logparser.exceptions.IncorrectParseDefinitionException;
 import com.adobe.campaign.tests.logparser.exceptions.ParseDefinitionImportExportException;
 import com.adobe.campaign.tests.logparser.exceptions.StringParseException;
@@ -1312,6 +1309,12 @@ public class LogDataTest {
         assertThat("We should have found 2 files", l_foundFilePaths,
                 Matchers.containsInAnyOrder(Matchers.endsWith("dirA/simpleLog.log"),
                         Matchers.endsWith("dirB/simpleLog.log")));
+    }
+
+    /*************** #55 Exporting results **********************/
+    @Test
+    public void testExportData() {
+
     }
 
 }
