+++ src/test/java/com/adobe/campaign/tests/logparser/core/TestLogParsing.java
@@ -9,7 +9,7 @@
  *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.adobe.campaign.tests.logparser;
+package com.adobe.campaign.tests.logparser.core;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
@@ -24,12 +24,13 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import com.adobe.campaign.tests.logparser.core.ParseDefinition;
 import org.hamcrest.Matchers;
 import org.testng.annotations.Test;
 
-import com.adobe.campaign.tests.logparser.GenericEntry;
-import com.adobe.campaign.tests.logparser.ParseDefinitionEntry;
-import com.adobe.campaign.tests.logparser.StringParseFactory;
+import com.adobe.campaign.tests.logparser.core.GenericEntry;
+import com.adobe.campaign.tests.logparser.core.ParseDefinitionEntry;
+import com.adobe.campaign.tests.logparser.core.StringParseFactory;
 import com.adobe.campaign.tests.logparser.exceptions.StringParseException;
 
 public class TestLogParsing {
